feat(api): move adminMember serializer to src
